requirements.txt
requests>=2.31.0
pyyaml>=6.0.1
maxminddb-geolite2>=2018.703
gitpython>=3.1.40 